Add official ARMhf and ARM64 Linux github actions
* Changed CI build and workflow. Details,
Added linux arm build. Linux arm build supports wayland (musescore#24111), x86_64 build does not yet.
Backported arm related: musescore#15923, #18120 (part of #18214), #20434, #24228 (#24675, #24696), #27785.
Added arm qt515webengine package (setup.sh L248) instead of mutating online save function considering current musescore.com server status musescore#27386.
The musescore#11913 (Qt5 QtQuick Controls 2 hack) had been added and removed in musescore 4 (Qt5 phase out). It is not backported as not needed on musescore 3.
Changed CI cpu from 2 to 4.
Fixed vtest comparison Proposed change build uses Base's build tools.
Fixed workflow regressions schedule build is not nightly (affects filename only), and BuildAll ignores default values.
Enabled vtest in Publishdev workflow.
---------

+++ build/ci/linux/tools/make_appimage.sh
@@ -1,19 +1,37 @@
 #!/usr/bin/env bash
+echo "make_appimage.sh start"
+trap 'echo making_appimage.sh failed; exit 1' ERR
 
 INSTALL_DIR="$1" # MuseScore was installed here
 APPIMAGE_NAME="$2" # name for AppImage file (created outside $INSTALL_DIR)
+PACKARCH="$3" # architecture (x86_64, aarch64, armv7l)
 
 if [ -z "$INSTALL_DIR" ]; then echo "error: not set INSTALL_DIR"; exit 1; fi
 if [ -z "$APPIMAGE_NAME" ]; then echo "error: not set APPIMAGE_NAME"; exit 1; fi
+if [ -z "$PACKARCH" ]; then 
+  PACKARCH="x86_64"
+elif [ "$PACKARCH" == "armv7l" ]; then
+  PACKARCH="armhf"
+fi
 
-##########################################################################
-# INSTALL APPIMAGETOOL AND LINUXDEPLOY
-##########################################################################
+HERE="$(cd "$(dirname "$0")" && pwd)"
+ORIGIN_DIR=${PWD}
+BUILD_TOOLS=$HOME/build_tools
+
+mkdir -p $BUILD_TOOLS
+
+echo "############################## INSTALL APPIMAGETOOL AND LINUXDEPLOY ##############################"
 
 function download_github_release()
 {
   local -r repo_slug="$1" release_tag="$2" file="$3"
-  wget -q --show-progress "https://github.com/${repo_slug}/releases/download/${release_tag}/${file}"
+  if [[ "${release_tag}" == "latest" ]]; then
+    local -r url="https://github.com/${repo_slug}/releases/latest/download/${file}"
+  else
+    local -r url="https://github.com/${repo_slug}/releases/download/${release_tag}/${file}"
+  fi
+  # use curl instead of wget which fails on armhf
+  curl "${url}" -O -L
   chmod +x "${file}"
 }
 
@@ -22,7 +40,12 @@ function extract_appimage()
   # Extract AppImage so we can run it without having to install FUSE
   local -r appimage="$1" binary_name="$2"
   local -r appdir="${appimage%.AppImage}.AppDir"
-  "./${appimage}" --appimage-extract >/dev/null # dest folder "squashfs-root"
+  # run appimage in docker container with QEMU emulation directly since binfmt fails
+  if [[ "$PACKARCH" == armhf ]]; then
+    /usr/bin/qemu-arm-static "./${appimage}" --appimage-extract >/dev/null # dest folder "squashfs-root"
+  else
+    "./${appimage}" --appimage-extract >/dev/null # dest folder "squashfs-root"
+  fi
   mv squashfs-root "${appdir}" # rename folder to avoid collisions
   ln -s "${appdir}/AppRun" "${binary_name}" # symlink for convenience
   rm -f "${appimage}"
@@ -31,48 +54,56 @@ function extract_appimage()
 function download_appimage_release()
 {
   local -r github_repo_slug="$1" binary_name="$2" tag="$3"
-  local -r appimage="${binary_name}-x86_64.AppImage"
+  local -r appimage="${binary_name}-${PACKARCH}.AppImage"
   download_github_release "${github_repo_slug}" "${tag}" "${appimage}"
   extract_appimage "${appimage}" "${binary_name}"
+  # mv "${appimage}" "${binary_name}" # use this instead of the previous line for the static runtime AppImage
 }
 
-if [[ ! -d "appimagetool" ]]; then
-  mkdir appimagetool
-  cd appimagetool
-  # `12` and not `continuous` because see https://github.com/AppImage/AppImageKit/issues/1060
-  download_appimage_release AppImage/AppImageKit appimagetool 12
-  cd ..
+if [[ ! -d $BUILD_TOOLS/appimagetool ]]; then
+  mkdir $BUILD_TOOLS/appimagetool
+  cd $BUILD_TOOLS/appimagetool
+  download_appimage_release AppImage/AppImageKit appimagetool continuous # use AppImage/appimagetool for the static runtime AppImage
+  cd $ORIGIN_DIR
 fi
-export PATH="${PWD%/}/appimagetool:${PATH}"
+export PATH="$BUILD_TOOLS/appimagetool:$PATH"
 appimagetool --version
-
-if [[ ! -d "appimageupdatetool" ]]; then
-  mkdir appimageupdatetool
-  cd appimageupdatetool
-  download_appimage_release AppImage/AppImageUpdate appimageupdatetool continuous
-  cd ..
-fi
-export PATH="${PWD%/}/appimageupdatetool:${PATH}"
-appimageupdatetool --version
 
 function download_linuxdeploy_component()
 {
   download_appimage_release "linuxdeploy/$1" "$1" continuous
 }
 
-if [[ ! -d "linuxdeploy" ]]; then
-  mkdir linuxdeploy
-  cd linuxdeploy
+if [[ ! -f $BUILD_TOOLS/linuxdeploy/linuxdeploy ]]; then
+  mkdir -p $BUILD_TOOLS/linuxdeploy
+  cd $BUILD_TOOLS/linuxdeploy
   download_linuxdeploy_component linuxdeploy
-  download_linuxdeploy_component linuxdeploy-plugin-qt
-  cd ..
+  cd $ORIGIN_DIR
 fi
-export PATH="${PWD%/}/linuxdeploy:${PATH}"
+if [[ ! -f $BUILD_TOOLS/linuxdeploy/linuxdeploy-plugin-qt ]]; then
+  mkdir -p $BUILD_TOOLS/linuxdeploy
+  cd $BUILD_TOOLS/linuxdeploy
+  download_linuxdeploy_component linuxdeploy-plugin-qt
+  cd $ORIGIN_DIR
+fi
+export PATH="$BUILD_TOOLS/linuxdeploy:$PATH"
 linuxdeploy --list-plugins
 
-##########################################################################
-# BUNDLE DEPENDENCIES INTO APPDIR
-##########################################################################
+if [[ ! -d $BUILD_TOOLS/appimageupdatetool ]]; then
+  mkdir $BUILD_TOOLS/appimageupdatetool
+  cd $BUILD_TOOLS/appimageupdatetool
+  download_appimage_release AppImage/AppImageUpdate appimageupdatetool continuous
+  cd $ORIGIN_DIR
+fi
+if [[ "${UPDATE_INFORMATION}" ]]; then
+  export PATH="$BUILD_TOOLS/appimageupdatetool:$PATH"
+
+  # `appimageupdatetool`'s `AppRun` script gets confused when called via a symlink.
+  # Resolve the symlink here to avoid this issue.
+  $(readlink -f "$(which appimageupdatetool)") --version
+fi
+
+echo "############################## BUNDLE DEPENDENCIES INTO APPDIR ##############################"
 
 cd "$(dirname "${INSTALL_DIR}")"
 appdir="$(basename "${INSTALL_DIR}")" # directory that will become the AppImage
@@ -87,8 +118,17 @@ mv "${appdir}/bin/findlib" "${appdir}/../findlib"
 qt_sql_drivers_path="${QT_PATH}/plugins/sqldrivers"
 qt_sql_drivers_tmp="/tmp/qtsqldrivers"
 mkdir -p "$qt_sql_drivers_tmp"
-mv "${qt_sql_drivers_path}/libqsqlmysql.so" "${qt_sql_drivers_tmp}/libqsqlmysql.so"
-mv "${qt_sql_drivers_path}/libqsqlpsql.so" "${qt_sql_drivers_tmp}/libqsqlpsql.so"
+[ -f "${qt_sql_drivers_path}/libqsqlmysql.so" ] && mv "${qt_sql_drivers_path}/libqsqlmysql.so" "${qt_sql_drivers_tmp}/libqsqlmysql.so"
+[ -f "${qt_sql_drivers_path}/libqsqlpsql.so" ] && mv "${qt_sql_drivers_path}/libqsqlpsql.so" "${qt_sql_drivers_tmp}/libqsqlpsql.so"
+
+# Semicolon-separated list of platforms to deploy in addition to `libqxcb.so`.
+# Used by linuxdeploy-plugin-qt.
+if [[ "$PACKARCH" == x86_64 ]]; then # current x86_64 qt5.9 does not contain wayland, setup.sh L200
+  export EXTRA_PLATFORM_PLUGINS=""
+else
+  export EXTRA_PLATFORM_PLUGINS="libqwayland-egl.so;libqwayland-generic.so" # # todo: arm related, the current qt5.9 package does not support wayland, see setup.sh L200
+  # libqoffscreen.so not backported, it fixes musescore 4 headless regression https://github.com/musescore/MuseScore/issues/17247
+fi
 
 # Colon-separated list of root directories containing QML files.
 # Needed for linuxdeploy-plugin-qt to scan for QML imports.
@@ -98,18 +138,21 @@ export QML_SOURCES_PATHS=./
 linuxdeploy --appdir "${appdir}" # adds all shared library dependencies
 linuxdeploy-plugin-qt --appdir "${appdir}" # adds all Qt dependencies
 
-unset QML_SOURCES_PATHS
+# (unsure backport or not) file dialog https://github.com/musescore/MuseScore/issues/10836 The system must be used
+# if [ -f ${appdir}/lib/libglib-2.0.so.0 ]; then
+#   rm -f ${appdir}/lib/libglib-2.0.so.0 
+# fi
 
-# In case this container is reused multiple times, return the moved libraries back
-mv "${qt_sql_drivers_tmp}/libqsqlmysql.so" "${qt_sql_drivers_path}/libqsqlmysql.so"
-mv "${qt_sql_drivers_tmp}/libqsqlpsql.so" "${qt_sql_drivers_path}/libqsqlpsql.so"
+unset QML_SOURCES_PATHS EXTRA_PLATFORM_PLUGINS
+
+# Return the moved libraries back
+[ -f "${qt_sql_drivers_tmp}/libqsqlmysql.so" ] && mv "${qt_sql_drivers_tmp}/libqsqlmysql.so" "${qt_sql_drivers_path}/libqsqlmysql.so"
+[ -f "${qt_sql_drivers_tmp}/libqsqlpsql.so" ] && mv "${qt_sql_drivers_tmp}/libqsqlpsql.so" "${qt_sql_drivers_path}/libqsqlpsql.so"
 
 # Put the non-RUNPATH binaries back
 mv "${appdir}/../findlib" "${appdir}/bin/findlib"
 
-##########################################################################
-# BUNDLE REMAINING DEPENDENCIES MANUALLY
-##########################################################################
+echo "############################## BUNDLE REMAINING DEPENDENCIES MANUALLY ##############################"
 
 function find_library()
 {
@@ -137,24 +180,47 @@ function fallback_library()
 # Report new additions at https://github.com/linuxdeploy/linuxdeploy/issues
 # or https://github.com/linuxdeploy/linuxdeploy-plugin-qt/issues for Qt libs.
 unwanted_files=(
-  # none
+  # https://github.com/musescore/MuseScore/issues/24068#issuecomment-2297823192
+  lib/libwayland-client.so.0
 )
 
 # ADDITIONAL QT COMPONENTS
 # linuxdeploy-plugin-qt may have missed some Qt files or folders that we need.
 # List them here using paths relative to the Qt root directory. Report new
 # additions at https://github.com/linuxdeploy/linuxdeploy-plugin-qt/issues
-additional_qt_components=(
-  /plugins/printsupport/libcupsprintersupport.so
-)
+if [[ "$PACKARCH" == "x86_64" ]]; then # current x86_64 qt5.9 does not contain wayland, setup.sh L200
+  additional_qt_components=(
+    plugins/printsupport/libcupsprintersupport.so
+    
+    # (unsure backport or not) file dialog https://github.com/musescore/MuseScore/issues/10836 # At an unknown point in time, the libqgtk3 plugin stopped being deployed
+    # plugins/platformthemes/libqgtk3.so
+  )
+else
+  additional_qt_components=(
+    plugins/printsupport/libcupsprintersupport.so
+
+    # (unsure backport or not) file dialog https://github.com/musescore/MuseScore/issues/10836 # At an unknown point in time, the libqgtk3 plugin stopped being deployed
+    # plugins/platformthemes/libqgtk3.so
+
+    # Wayland support (run with QT_QPA_PLATFORM=wayland to use)
+    plugins/wayland-decoration-client
+    plugins/wayland-graphics-integration-client
+    plugins/wayland-shell-integration
+  )
+fi
 
 # ADDITIONAL LIBRARIES
 # linuxdeploy may have missed some libraries that we need
 # Report new additions at https://github.com/linuxdeploy/linuxdeploy/issues
-additional_libraries=(
-  libssl.so.1.0.0    # OpenSSL (for Save Online)
-  libcrypto.so.1.0.0 # OpenSSL (for Save Online)
-)
+if [[ "$PACKARCH" == "x86_64" ]]; then
+  additional_libraries=(
+    libssl.so.1.1    # OpenSSL (for Save Online)
+    libcrypto.so.1.1 # OpenSSL (for Save Online). openssl missing after github ubuntu update https://github.com/musescore/MuseScore/issues/18120#issuecomment-1604116777
+  )
+else
+  additional_libraries=()
+fi
+
 
 # FALLBACK LIBRARIES
 # These get bundled in the AppImage, but are only loaded if the user does not
@@ -164,27 +230,48 @@ additional_libraries=(
 # startup if the library was not found. The fallback library may not provide
 # the full functionality of the system version, but it does avoid the crash.
 # Report new additions at https://github.com/linuxdeploy/linuxdeploy/issues
-fallback_libraries=(
-  libjack.so.0 # https://github.com/LMMS/lmms/pull/3958
-)
+if [[ "$PACKARCH" == "x86_64" ]]; then
+  fallback_libraries=(
+    libjack.so.0 # https://github.com/LMMS/lmms/pull/3958
+  )
+else
+  fallback_libraries=(
+    libjack.so.0 # https://github.com/LMMS/lmms/pull/3958
+    libOpenGL.so.0 # fixes ARM startup bug https://github.com/musescore/MuseScore/issues/24228 , qt5 and qt6 https://bugreports.qt.io/browse/QTBUG-89754, libopengl-dev in setup.sh
+  )
+fi
 
 # PREVIOUSLY EXTRACTED APPIMAGES
 # These include their own dependencies. We bundle them uncompressed to avoid
 # creating a double layer of compression (AppImage inside AppImage).
-extracted_appimages=(
-  appimageupdatetool
-)
+if [[ "${UPDATE_INFORMATION}" ]]; then
+  extracted_appimages=(
+    appimageupdatetool
+  )
+else
+  extracted_appimages=(
+    # none
+  )
+fi
 
 for file in "${unwanted_files[@]}"; do
   rm -rf "${appdir}/${file}"
 done
 
 for file in "${additional_qt_components[@]}"; do
-  mkdir -p "${appdir}/$(dirname "${file}")"
-  cp -L "${QT_PATH}/${file}" "${appdir}/${file}"
+  if [ -f "${appdir}/${file}" ]; then
+    echo "Warning: ${file} was already deployed. Skipping."
+    continue
+  fi
+    mkdir -p "${appdir}/$(dirname "${file}")"
+    cp -Lr "${QT_PATH}/${file}" "${appdir}/${file}"
 done
 
 for lib in "${additional_libraries[@]}"; do
+  if [ -f "${appdir}/lib/${lib}" ]; then
+    echo "Warning: ${file} was already deployed. Skipping."
+    continue
+  fi
   full_path="$(find_library "${lib}")"
   cp -L "${full_path}" "${appdir}/lib/${lib}"
 done
@@ -203,7 +290,13 @@ for name in "${extracted_appimages[@]}"; do
   extracted_appdir_path="$(dirname "${apprun}")"
   extracted_appdir_name="$(basename "${extracted_appdir_path}")"
   cp -r "${extracted_appdir_path}" "${appdir}/"
-  ln -s "../${extracted_appdir_name}/AppRun" "${appdir}/bin/${name}"
+  cat >"${appdir}/bin/${name}" <<EOF
+#!/bin/sh
+unset APPDIR APPIMAGE # clear outer values before running inner AppImage
+HERE="\$(dirname "\$(readlink -f "\$0")")"
+exec "\${HERE}/../${extracted_appdir_name}/AppRun" "\$@"
+EOF
+  chmod +x "${appdir}/bin/${name}"
 done
 
 # METHOD OF LAST RESORT
@@ -238,14 +331,13 @@ for file in "${libnss3_files[@]}"; do
   rm -f "${appdir}/lib/$(basename "${file}")" # in case it was already packaged by linuxdeploy
 done
 
-##########################################################################
-# TURN APPDIR INTO AN APPIMAGE
-##########################################################################
+
+echo "############################## TURN APPDIR INTO AN APPIMAGE ##############################"
 
 appimage="${APPIMAGE_NAME}" # name to use for AppImage file
 
 appimagetool_args=( # array
-  # none
+  --no-appstream # do not check upstream metadata
   )
 
 created_files=(
@@ -270,14 +362,4 @@ fi
 # create AppImage
 appimagetool "${appimagetool_args[@]}" "${appdir}" "${appimage}"
 
-# We are running as root in the Docker image so all created files belong to
-# root. Allow non-root users outside the Docker image to access these files.
-chmod a+rwx "${created_files[@]}"
-parent_dir="${PWD}"
-while [[ "$(dirname "${parent_dir}")" != "${parent_dir}" ]]; do
-  [[ "$parent_dir" == "/" ]] && break
-  chmod a+rwx "$parent_dir"
-  parent_dir="$(dirname "$parent_dir")"
-done
-
-echo "Making AppImage finished"
+echo "make_appimage.sh ended"
